Quick fix travis oom with test race
